Added file locking for archive files, so theoretically the syncing app can be run in parallel.

+++ src/archive.rs
@@ -1,12 +1,14 @@
 use std::io;
 use std::convert::From;
 use std::hash::Hasher;
-use std::fs::{File, create_dir, remove_file};
+use std::fs::{File, create_dir_all, remove_file};
 use std::path::{Path, PathBuf};
 use bincode::serde::{serialize_into, deserialize_from, DeserializeError, SerializeError};
 use bincode::SizeLimit;
 use util::hash_single;
 use byteorder::{Error as ByteorderError, WriteBytesExt, ReadBytesExt, LittleEndian};
+//use file_lock::{Error, Lock, LockKind, AccessMode};
+use fs2::FileExt;
 
 use state::{HashedPath, ArchiveEntry, ArchiveEntryPerReplica};
 use util::FnvHashMap;
@@ -22,19 +24,17 @@ impl Archive {
     pub fn new(directory: PathBuf) -> Result<Self, io::Error> {
         // creates the archive directory
         if !directory.exists() {
-            try!(create_dir(&directory));
+            try!(create_dir_all(&directory));
         }
         Ok(Archive { directory: directory })
-    }
-
-    fn file_for_directory(&self, directory: &Path) -> PathBuf {
-        self.directory.join(hash_single(directory).to_string())
     }
 
     /// Returns an `ArchiveFile` struct which abstracts over operations on a single archive file.
     /// Remember each 'file' in the archive represents an entire directory (not recursive) in the replicas.
     pub fn for_directory(&self, directory: &Path) -> ArchiveFile {
-        return ArchiveFile { path: self.file_for_directory(directory) }
+        let path = self.directory.join(hash_single(directory).to_string());
+
+        ArchiveFile::new(path)
     }
 }
 
@@ -45,10 +45,38 @@ pub struct ArchiveFile {
 }
 
 impl ArchiveFile {
+
+    /// Creates a new wrapper around the given archive file.
+    /// Through doing so this will acquire (or wait for) a lock,
+    /// ensuring that multiple threads/processes aren't writing to the same archive file.
+    pub fn new(path: PathBuf) -> ArchiveFile {
+        /*let lock_path = path.with_extension("lock");
+        let lock_file = OpenOptions::new().create(true).open(lock_path).unwrap();
+        let lock = Lock::new(lock_file.as_raw_fd());
+        match lock.lock(LockKind::NonBlocking, AccessMode::Write) {
+            Ok(_) => (),
+            Err(Error::Errno(i))
+              => println!("Got filesystem error while locking {}", i),
+        }*/
+
+        if path.exists() {
+            let f = File::open(path.clone()).unwrap();
+            trace!("Acquiring lock for {:?}", path);
+            f.lock_exclusive().unwrap();
+            trace!("Acquired lock");
+        }
+
+        ArchiveFile { path: path }
+    }
+
     /// Remove all entries from this file.
     /// This just slightly more efficient than writing an empty Vec.
     pub fn remove_all(&self) -> Result<(), io::Error> {
-        remove_file(self.path.clone())
+        if self.path.exists() {
+            remove_file(self.path.clone())
+        } else {
+            Ok(())
+        }
     }
 
     /// Reads the archive entries into a Vec
@@ -74,10 +102,10 @@ impl ArchiveFile {
 
         if entries.is_empty() {
             if self.path.exists() {
-                debug!("Removing archive file {:?} (all entries gone)", self.path);
+                debug!("Removing archive file {:?} (because entries are empty)", self.path);
                 try!(remove_file(self.path.clone()));
             } else {
-                debug!("Archive file {:?} doesn't exist (all entries gone)", self.path);
+                debug!("Tried to remove archive file {:?} (because entries are empty), but it doesn't exist ", self.path);
             }
         } else {
             debug!("Writing archive file {:?}", self.path);
@@ -86,6 +114,18 @@ impl ArchiveFile {
         }
 
         Ok(())
+    }
+}
+
+impl Drop for ArchiveFile {
+    fn drop(&mut self) {
+        if self.path.exists() {
+            let f = File::open(self.path.clone()).unwrap();
+            trace!("Unlocking archive file {:?}", self.path);
+            f.unlock().unwrap();
+            trace!("Unlocked");
+        }
+        //self.lock.unlock().unwrap();
     }
 }
 
@@ -128,8 +168,8 @@ impl ArchiveEntries {
     pub fn insert(&mut self, path: &Path, entries: Vec<ArchiveEntryPerReplica>) {
         let hashed_path = hash_single(path);
 
-        // warn because it means we are potentially being inefficient
-        warn!("Inserting data into archive for path {:?} (hashed: {})\n", path, hashed_path);
+        // this means we are potentially being inefficient
+        info!("Inserting data into archive for path {:?} (hashed: {})", path, hashed_path);
         self.entries.insert(hashed_path, entries);
         self.dirty = true;
     }
@@ -148,7 +188,7 @@ fn remove_deleted_entries(entries: &mut Vec<ArchiveEntry>) {
         }
 
         if !keep {
-            info!("Removing stale entry before writing to archive");
+            info!("Removing empty entry before writing to archive");
         }
         keep
     });
